rgw: pick the last ip in x-forwarded-for chain
Currently, when rgw_remote_addr_param is set to HTTP_X_FORWARDED_FOR,
it will pick the first IP from the chain. As described here
(https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Forwarded-For)
it needs to pick the last one otherwise it can be manipulated by
the client and cause access to a bucket (which is protected by
a bucket policy based on aws:SourceIP).

The introduced option `rgw_trusted_proxy_count` allows the user to
identify the index which the IP should be picked by the chain of
X-Forwarded-For header and to stay with the old evaluation.

Fixes: https://tracker.ceph.com/issues/64489
Signed-off-by: Seena Fallah <seenafallah@gmail.com>