librbd: propagate blkdiscards in mirroring scenarios
Unlike in an rbd deep copy, blkdiscards were not being propagated to the
non-primary rbd mirror image because ObjectCopyRequest::compute_zero_ops
skipped the the zeroed extents when performing incremental snapshot copies.
This fix adds the zeroed object extents to the m_snapshot_sparse_bufferlist
in incremental copy ops so that a remove op is sent to the non-primary image.
It also updates the object map for blkdiscard ops using a newly added
optimization in the object map update check.

Fixes: https://tracker.ceph.com/issues/58852

Signed-off-by: N Balachandran <nibalach@redhat.com>