src/mon/ElectionLogic: Improve victory logic for connectivity strategy
Problem:

When there are 3 MONs in 3 different hosts [mon.a, mon.b, mon.c (tiebreaker)]
and there is a netsplit scenario between mon.a and mon.b
while mon.a is still able to connect to mon.c and mon.b still
able to connect to mon.c, lets say mon.a has the best connection score
it still has to wait until the end of an election period before
sending out a victory message because the current logic states
that monitor needs to wait for all acks == quorum.size before delcaring victory
or if acks >= quorum.size/2 when election ends. Sometimes, the later
case turns out to create a never ending election because in this example,
mon.b would also reach an end of election period and starts a new election,
bumping the epoch to higher than mon.c, then propose to mon.c which then
the incoming victory message from mon.a gets dropped because mon.c
bumps its epoch due to mon.b's proposal.

Solution:

Change the declare victory logic in Connectivity Strategy to be
If we recieved >= quorum.size/2 and the remaining monitors matches
the dead_pinging set we tracked (meaning we declared
them dead), then we can go ahead and delcare ourself victory.

Also, changed the logic of how we handle victory message
where if the mon recieves 2 consecutive victory messages
we drop the later one because we assume who ever sends it
first probably is the true winner.

Signed-off-by: Kamoltat <ksirivad@redhat.com>