src/common/mempool.cc: Adaptive caching of sched_getcpu() to select shard
Improve the performance of mempool sharding on systems with sched_getcpu()

The idea is to assign threads to shards based on which CPU core they are
executing on using sched_getcpu() to select the shard. All the threads
executing on the same CPU core share the same shard. If each CPU core has its
own shard there should be no cache line contention.

There is a cost to repeatedly calling sched_getcpu() so cache this in a thread
local variable. Except for Crimson WITH_SEASTAR and !WITH_ALIEN builds, threads
are not bound (have affinity) to a single CPU core so there is a risk that the
cached choice of shard will become invalid if the kernel scheduler decides to
migrate a thread to another core. This will then result in poor performance due
to CPU cache lines ping ponging between cores (correctness is maintained because
shards still use atomics). Mitigate this risk by periodically checking which
core the thread is running on and updating the cached value. Adapt how
frequently the code performs this check based on whether the thread has
migrated.

On a busy system (where CPU resource is liable to become a bottleneck) it is
probably that a thread will make multiple calls to mempool before a context
switch occurs. Modern kernel schedulers understand there is a cost to migrating
a thread so will try to avoid this unless another core is much less heavily
loaded.

For a Crimson WITH_SEASTAR and !WITH_ALIEN build the threads will not migrate
between cores so the periodic checking of core assignment will become so rare
that it has negligible impact on performance.

Signed-off-by: Bill Scales <bill_scales@uk.ibm.com>