multi-mds/link: Load referent inodes if path2's depth is 0
If the link doesn't come with source path and only targeti
is passed, the referent inodes were not loaded. This is
fixed with this.

Also use projected_inode of target inode for loading referent inodes.
The link request could return with early_reply without the projected
inode being popped. The immediate next requests on the same link wouldn't
find the added referent inode in the list. This would return stale list
of snaprealms to the client.

Signed-off-by: Kotresh HR <khiremat@redhat.com>