mgr/volumes: handling dangling symlinks gracefully
If we come across situations where we encounter dangling
symlinks for clones due to any reason,like older versions
may have produced dangling clone symlinks which remained
post-upgrade, it needs to be handled gracefully by deleting
it, presence of which may not allow deletion of the snapshot.
We are now cleaning those up in the snapshot info command.

Fixes: https://tracker.ceph.com/issues/58090
Signed-off-by: Neeraj Pratap Singh <neesingh@redhat.com>