mon/OSDMonitor: fix rmsnap command
```
There are 2 ways to remove pool snaps, rados tool or mon command (ceph osd pool rmsnap).
It seems that the monitor command is not reporting the actual removal via new_removed_snaps which is later proceed in OSDMap::apply_incremental.
This will result in a clone object leakage since the snap id won't be marked as purged (and won't be trimmed).
```

Fixes: https://tracker.ceph.com/issues/64646

Signed-off-by: Matan Breizman <mbreizma@redhat.com>