qa/crontab: correct script paths and environment
At some point the links to the shell scripts in ceph.git were broken in the
$HOME for [1]. Unless a run was done manually with `teuthology-suite` in the
crontab, the job was basically skipped. This is probably one of the reasons
nightlies fell out of use.

I've updated the home directory according to the document (comments) in this
change. The teuthology user now has persistent clones of ceph.git and
teuthology.git. The clones are updated daily by this same crontab.

Instead of using a link in teuthology's $HOME/bin to the scripts used in this
crontab, we just have a cron variable referencing where the script should be in
the ceph.git/teuthology.git clone. Adding to this, the .bash_environment file
sources the virtualenv activate script instead of adding the teuthology binary
directory to its $PATH.

I've updated the hour for these jobs to actually be done "nightly". The first
set of jobs will be scheduled around 4pm EST. Additionally, it was necessary to
include --force-priority as some jobs are below the priority thresholds.

[1] teuthology@teuthology.front.sepia.ceph.com.

Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>