rgw: allow s3 bucket listing to return zero entries
When a bucket index shard had many contiguous delete markers,
a single bucket listing request could cause huge read
amplification on the backend because RGW needed to iteratively loop
through all of the delete markers in order to find at least one
listable entry to sort among other shards and return a response
to the client. In extreme cases, long latencies due to excessive
iteration led to client timeouts and client retries, which
compounded impact on cluster performance.

The AWS S3 specification does not make any promises
about the number of entries that will be returned
and claims that continuation tokens are obfuscated and not
real keys. This permits us to workaround the aforementioned
problem by returning zero keys when excessive iteration is
required, advancing the continuation token to a marker associated
with an un-listable entry, and letting the client drive continued
iteration with subsequent requests.

Fixes: https://tracker.ceph.com/issues/64185
Signed-off-by: Cory Snyder <csnyder@1111systems.com>