mon: Use minimum pg_num as 8 for pool creation if autoscale mode is on
Creating a new pool with 1 PG when autoscale mode is ON results
in splitting this PG into many on a single OSD causing OSD
to hit max_pgs_per_osd.

Use minimum pg_num as 8 to avoid a situation where an OSD
hits max_pgs_per_osd and prevents a new PG creation.

Fixes: https://tracker.ceph.com/issues/64938

Signed-off-by: Prashant D <pdhange@redhat.com>