osd/scrubber: Add a function to calculate scrub cost for mClock
With mClock scheduler, we need the cost of an operation to represent the size of an operation. This function will calculate the average object size of the PG that needs to be scrubbed and return it as the scrub cost.

Signed-off-by: Aishwarya Mathuria <amathuri@redhat.com>
(cherry picked from commit c269f352e538b795858b963644a05a65a04b6a8e)