crimson/osd/pg: use pgmeta_oid for snapmapper
Using a new object for snapmapper is tricky.  The only existing object
like this is the pg meta object, which is special-cased in a variety of
places through the osd.  It also happens to have the nice property that
it's not expressible through a normal rados operation -- rados objects
are not allowed to have empty names.

This snapmapper object, on the other hand, could actually be the target
of a user read or write (though it's unlikely to happen by accident).
In order to do this correctly, the existing special cases for the pgmeta
object would need to be generalized and the normal user IO path would
need to disallow mutations (and maybe reads since these objects wouldn't
follow normal rules).

Instead, let's just stick these keys into the pgmeta object and save
ourselves a lot of work.

Fixes: https://tracker.ceph.com/issues/64975
Signed-off-by: Samuel Just <sjust@redhat.com>