script: allow use of run_tox.sh with default tox envlist
Add a "magic" --tox-envs value to force the run_tox.sh script to
defer to the tox.ini file's envlist rather than specifying a
custom env list. Pass `--tox-envs __tox_defaults__` to the script
to use whatever the tox.ini defaults are.

Adding a magic value like this avoids the need to make breaking changes
(or any changes) to the current cmake "code" used to invoke the
tox tests from make check. It's a little hacky but the funny looking
value should make it clear that it's special. In particular it avoids
adding new conditionals to the cmake files.

Signed-off-by: John Mulligan <jmulligan@redhat.com>