qa/cephfs: don't use run_cluster_cmd()
The method was deleted in favour of run_ceph_cmd() and
get_ceph_cmd_stdout() in commit
35a0d84972d691535533ffeb1635e5f94d808575.

This commit is not a backport but a new commit to fix the code present
on Reef branch but not present in release branches after Reef. At least
some of these issues were exposed during QA done for the backport.

Signed-off-by: Rishabh Dave <ridave@redhat.com>