qa/cephfs: in test_volumes.py, don't use deleted helper methods
_generate_random_group_name() and _generate_random_subvolume_name() were
deleted in one of the commit from this series of commits. And PR #54207
adds calls to these helper methods. This wasn't the case for the main
branch version of this series of commits. This is why this issue is
being fixed in Reef branch directly through a fresh/new commit instead
of backporting a commit.

Signed-off-by: Rishabh Dave <ridave@redhat.com>