DNM: qa/cephfs: running proc launched by a cmd in bg
It should've been possible to put the the process launched by tar
command to run in background, while tests runs following commands in
foreground by passing "wait=False" to run_shell() method.

However, this doesn't work in case of tar command. tar command is
launched when subprocess.Popen() command is instantiated (let's call
this instance "subproc" hereafter) but the process launched by the
command halts within 3-4 seconds. It takes subproc.communicate() method
to be executed to get the process running again. And then it runs till
completion.

Th is not the case of processes launched for simpler commands like "echo
1 > a; sleep 5; echo 2 >> a". Most likely this happens for the process
launched by tar because large amount of data is written to stdout/stderr
by process. Which is calling subproc.communicate() makes it work, it
reads the data from the buffer for stdout/stderr which when full halts
the process executions.

Signed-off-by: Rishabh Dave <ridave@redhat.com>