mgr/rgw: add support to modify zonegroup parameters
This is being done with `radosgw-admin zonegroup set`
rather than `radosgw-admin zonegroup modify` as I don't
think the hostnames parameter (which is the primary focus
for this specific change) can be set using the modify
command. The nice bit about that is it should in theory
make it easy to extend this to allow setting other parameters
to be modified in the zonegroup in the future.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 5f7c6d58cc90ff139c1f8f8b950ef37d8597f92f)

Conflicts:
	src/python-common/ceph/deployment/service_spec.py