cephadm: allow passing custom prometheus alerts to bootstrap
This can't be done with the --config flag as the thing
that needs to be changed is in the config-key store rather
than being a config option. I don't think we have any
option other than a new flag here. Makes use of the
`ceph orch prometheus set-custom-alerts` command being
added in another commit

Signed-off-by: Adam King <adking@redhat.com>