qa/lsan.supp: suppress leak in ceph::ErasureCodePluginRegistry::load
`ceph::ErasureCodePluginRegistry::load()` is known to leak, as we
don't free the memory allocated by the ec plugins which are registered
in the `ErasureCodePluginRegistry` singleton. this is a known issue,
but since the `ErasureCodePluginRegistry` instance is a singleton.
we can live with it.

in this change, we add the rule to suppress the leak report from
LeakSanitizer. this rule also exist in qa/valgrind.supp.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>