mgr/dashboard: modernize REST API infra
The goal of this work is to adapt the REST API infrastructure to the
modern days, taking inspiration on modern frameworks like FastAPI.

The Ceph Dashboard API was conceived under Python 2.7, which lacked many
of the features available today that especially useful for generating
OpenAPI specs. Essentially, Python `typing` and type hints.

So goals for this work are:
* Remove the need of custom decorators (`Endpoint`, `DocEndpoint`,
  `ApiEndpoint`, etc), and infer documentation strings from docstrings,
  and data types from type hints.
* Improve the quality of the OpenAPI spec by adding return types,
  resposnes, models and others.
* Reduce the amount of boilerplate code required
* Increase the type-safety at the different stages of the code:
  * Automatically convert types like JSON Bool or Number to Python
    `bool` or `int` or `float`.

Signed-off-by: Ernesto Puerta <epuertat@redhat.com>