rgw: udpate options yaml file so LDAP uri isn't an invalid example
LDAP tries to bind the URI configuration option when RGW starts. The
default value is an example used to show the form of the URI and is
not itself valid. The default value is used, unless overrideen, and
can cause delays in start-up in some situations. The example is now
provided in the description and the default is the empty string.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 236f1dfa9ae507940a53301f31da7974456c82fe)

Conflicts:
    src/rgw/librgw.cc
        - changes here were in rgw_appmain.cc in main branch