mgr/cephadm: pass daemon's current image when reconfiguring
Important to note here is that a reconfig will rewrite
the config files of the daemon and restart it, but not
rewrite the unit files. This lead to a bug where the
grafana image we used between the quincy and squid release
used different UIDs internally, which caused us to rewrite
the config files as owned by a UID that worked with the
new image but did not work with the image still specified
in the unit.run file. This meant the grafana daemon was down
from a bit after the mgr upgrade until the end
of the upgrade when we redeploy all the monitoring images.

Fixes: https://tracker.ceph.com/issues/65234

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit 5922d20e05ef60414e72d11d0e947093d92dc288)

Conflicts:
	src/pybind/mgr/cephadm/tests/test_services.py