qa/cephadm: update images for test_cephadm workunit
This patch doesn't actually fix any tests, but looking
at https://tracker.ceph.com/issues/64208 has made me
realize we haven't been updating the images on here
at all. This patch just removes the octopus and pacific
image variables and puts in ones for reef and squid. I
think this is a reasonable way to try and maintain these,
as we're just keeping around the image for the last two
releases, which matches with what versions could one day
upgrade to what main will become (the "T" release as
of writing this)

Signed-off-by: Adam King <adking@redhat.com>