mgr/cephadm: additional debug logging for autotuner
This came from trying to debug behavior of the autotuner
in the upstream mailing list. The test case being added
was what that user was seeing. The debug logging being
added was useful in getting a full understanding of how
the autotuner got the result it did. Therefore, why not
add the logging to the actual codebase so we can make use
of it to debug autotuner issues in the future

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit a3f69e45cd802158012693937e65df5936e9f066)