Fix dependencies, so headers are built first
There's a race condition where on slow hosts, headers aren't
generated before starting the build process. The way to fix
this in a Debian package is to do:

override_dh_auto_build:
        dh_auto_build --buildsystem=cmake -- legacy-option-headers
        dh_auto_build --buildsystem=cmake

but this is a hack, and best is to get these dependencies
fixed in the build process directly. This is what this patch
is attempting to do.

Signed-off-by: Thomas Goirand <zigo@debian.org>
Co-Author: Xinhui Yang <cyan@cyano.uk>