librbd: data corruption on mirroring cloned images
Parent data is not synced to the non-primary after a copyup operation
on the primary image with snapshot based rbd-mirroring.
Writing to a non-existent object in an rbd clone image triggers a copyup
operation.The new  write data and the data from the parent are combined and
written to the new clone object and any existing snapshots on the clone are
updated with the parent data.
The rbd-mirror daemon uses deep_copy with subsets of snapshots. If a copyup
has modified the already synced older snapshot with the parent data, a
snapshot diff will only return the new data that was written to the clone. As
the deep_copy writes data to the dst object using rados calls instead of
rbd apis, a copyup is not performed and the parent data is lost.
The fix involves triggering a copyup on the object if the image is a clone.
This does not affect other deep_copy users as they copy the image and
all the existing snapshots as a whole.

Fixes: https://tracker.ceph.com/issues/61891

Signed-off-by: N Balachandran <nibalach@redhat.com>