osd/ECTransaction: Remove incorrect asserts in generate_transactions
Back in PR #11701 when EC Overwrites were added, there was significant churn in the ECTransaction code.  Several asserts were added in generate_transactions, but the code changed several times and it appears some of those asserts don't make sense in the final version of the PR.

This PR removes two of those asserts.  The first doesn't appear to be necessary given how the surrounding conditional block changed.   The second appears to be an incorrect assert that was left over from an earlier commit that no longer should be in place given the logic that was added to handle truncate (which roughly mirrors what we do in ReplicatedBackend).

It's possible the assert for obc and entry are also unnecessary, however I left those in place for now.

Fixes: https://tracker.ceph.com/issues/65509

Signed-off-by: Mark Nelson <mark.nelson@clyso.com>
(cherry picked from commit c6eb35b2b47a512779454ac91b1a0cd45c32b1bb)