osd: add watch ping timeout count in osd
For example, rbd send a watch ping to the header object
every 5 seconds to keep watch, if the primary OSD is unable
to receive the watch ping of the header object due to rbd
network interruption, this means that rbd's I/O has already been
hang. This way, we can quickly detect disconnection rbds on the osd,
and reflected in metrics.

Signed-off-by: Yite Gu <yitegu0@gmail.com>